requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pdf2image
pillow
torch
torchvision
transformers
pandas
openpyxl